requirements.txt
annotated-types==0.7.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
humanfriendly==10.0
idna==3.7
instagrapi==2.1.2
pillow==10.3.0
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pyreadline3==3.4.1
PySocks==1.7.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
